@@ -11,42 +11,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.0",
-    "@angular/common": "^8.0.0",
-    "@angular/compiler": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/forms": "^8.0.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
     "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@angular/router": "^8.0.0",
-    "bcrypt": "^3.0.6",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "bcrypt": "^3.0.7",
     "bootstrap-sass": "^3.4.1",
-    "ngx-bootstrap": "^4.2.0",
-    "rxjs": "^6.5.2",
-    "zone.js": "~0.9.1"
+    "ngx-bootstrap": "^5.2.0",
+    "rxjs": "^6.5.4",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.800.2",
-    "@angular/cli": "^8.0.2",
-    "@angular/compiler-cli": "^8.0.0",
-    "@angular/language-service": "^8.0.0",
-    "@types/jasmine": "~3.3.13",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~12.0.7",
-    "codelyzer": "~5.1.0",
-    "core-js": "^2.6.9",
-    "jasmine-core": "~3.4.0",
+    "@angular-devkit/build-angular": "^0.803.21",
+    "@angular/cli": "^8.3.21",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~13.1.4",
+    "codelyzer": "~5.2.1",
+    "core-js": "^3.6.1",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
+    "karma-jasmine": "~3.0.0",
+    "karma-jasmine-html-reporter": "^1.5.1",
     "protractor": "^5.4.2",
     "querystringify": "^2.1.1",
-    "ts-node": "~8.2.0",
-    "tslint": "~5.17.0",
-    "typescript": "^3.4.5"
+    "ts-node": "~8.5.4",
+    "tslint": "~5.20.1",
+    "typescript": "^3.5.3"
   }
 }
